requirements.txt
# AEGIS v3.0 Backend Requirements

# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# Scheduler
apscheduler==3.10.4

# Korean Stock Market Data
pykrx==1.0.46

# US Market Data
yfinance==0.2.35

# HTTP Clients
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# AI Models
anthropic==0.18.1
openai==1.12.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Telegram Bot
python-telegram-bot==20.7

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
rich==13.7.0
psutil==5.9.8
